requirements.txt
absl-py==1.4.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
cachetools==5.3.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.25.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.10.0
fonttools==4.39.3
ftfy==6.1.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==15.0.7
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
nltk==3.7
numpy==1.23.2
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==2.0.0+cu118
torch-tb-profiler==0.4.1
torchaudio==2.0.1+cu118
torchmetrics==0.11.0
torchvision==0.15.1+cu118
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
transformers==4.21.2
triton==2.0.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
